Fixes on dasherize and terminate the process when an error occur.
My dasherize function had a bug. Also, if some error occur, the children
processes keep running and the parent process "ate" the error.

Now things are smoother.

+++ lib/deadly_serious/engine/channel.rb
@@ -18,10 +18,12 @@ module DeadlySerious
       end
 
       def open_reader
+        fail %(File "#{@io_name}" not found) unless File.exist?(@io_name)
         open(@io_name, 'r')
       end
 
       def open_writer
+        fail %(File "#{@io_name}" not found) unless File.exist?(@io_name)
         open(@io_name, 'w')
       end
     end
